requirements.txt
aiohttp==3.7.4.post0
arrow==1.1.1
async-timeout==3.0.1
attrs==21.2.0
black==21.8b0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
discord==1.7.3
discord.py==1.7.3
idna==3.2
multidict==5.1.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.3.0
pydantic==1.8.2
python-dateutil==2.8.2
python-dotenv==0.19.0
regex==2021.8.28
requests==2.26.0
six==1.16.0
tomli==1.2.1
typing-extensions==3.10.0.2
urllib3==1.26.6
yarl==1.6.3
